+++ test/transforms/test_gdc.py
@@ -3,6 +3,7 @@ import torch
 from torch_geometric.data import Data
 from torch_geometric.testing import withPackage
 from torch_geometric.transforms import GDC
+from torch_geometric.transforms.gdc import sparsify_top_k
 from torch_geometric.utils import to_dense_adj
 
 
@@ -90,3 +91,16 @@ def test_gdc():
     assert torch.all(
         torch.isclose(col_sum, torch.tensor(1.0))
         | torch.isclose(col_sum, torch.tensor(0.0)))
+
+
+def test_sparsify_top_k():
+    edge_index = torch.tensor([[0, 1, 1, 1, 2, 2], [1, 0, 1, 2, 0, 1]])
+    edge_weight = torch.tensor([1, 1, 0, 2, 1, 2], dtype=torch.float)
+
+    out = sparsify_top_k(edge_index, edge_weight, k=1)
+    assert out[0].tolist() == [[0, 1, 2], [1, 2, 1]]
+    assert out[1].tolist() == [1, 2, 2]
+
+    out = sparsify_top_k(edge_index, edge_weight, k=2)
+    assert out[0].tolist() == [[0, 1, 1, 2, 2], [1, 2, 0, 1, 0]]
+    assert out[1].tolist() == [1, 2, 1, 2, 1]