+++ torch_geometric/nn/aggr/__init__.py
@@ -1,7 +1,8 @@
-from .base import Aggregation
+from .base import Aggregation, WeightedAggregation
 from .multi import MultiAggregation
 from .basic import (
     MeanAggregation,
+    WeightedMeanAggregation,
     SumAggregation,
     MaxAggregation,
     MinAggregation,
@@ -23,13 +24,20 @@ from .attention import AttentionalAggregation
 from .mlp import MLPAggregation
 from .deep_sets import DeepSetsAggregation
 from .set_transformer import SetTransformerAggregation
+from .soft_median import (
+    WeightedMedianAggregation,
+    WeightedQuantileAggregation,
+    SoftMedianAggregation,
+)
 from .lcm import LCMAggregation
 
 __all__ = classes = [
     'Aggregation',
     'MultiAggregation',
+    'WeightedAggregation',
     'SumAggregation',
     'MeanAggregation',
+    'WeightedMeanAggregation',
     'MaxAggregation',
     'MinAggregation',
     'MulAggregation',
@@ -50,5 +58,8 @@ __all__ = classes = [
     'MLPAggregation',
     'DeepSetsAggregation',
     'SetTransformerAggregation',
+    'WeightedMedianAggregation',
+    'WeightedQuantileAggregation',
+    'SoftMedianAggregation',
     'LCMAggregation',
 ]