+++ torch_geometric/nn/aggr/basic.py
@@ -5,7 +5,7 @@ import torch
 from torch import Tensor
 from torch.nn import Parameter
 
-from torch_geometric.nn.aggr import Aggregation
+from torch_geometric.nn.aggr import Aggregation, WeightedAggregation
 from torch_geometric.utils import softmax
 
 
@@ -33,6 +33,23 @@ class MeanAggregation(Aggregation):
                 ptr: Optional[Tensor] = None, dim_size: Optional[int] = None,
                 dim: int = -2) -> Tensor:
         return self.reduce(x, index, ptr, dim_size, dim, reduce='mean')
+
+
+class WeightedMeanAggregation(WeightedAggregation):
+    r"""An aggregation operator that computes the weighted average of features
+    across a set of elements
+
+    .. math::
+        \mathrm{mean}(\mathcal{X}, \mathbf{w}) = \frac{1}{|\mathcal{X}|}
+        \sum_{\mathbf{x}_i \in \mathcal{X}} w_i \mathbf{x}_i.
+    """
+    def forward(self, x: Tensor, index: Optional[Tensor] = None,
+                weight: Optional[Tensor] = None, ptr: Optional[Tensor] = None,
+                dim_size: Optional[int] = None, dim: int = -2) -> Tensor:
+        if weight is None:
+            return self.reduce(x, index, ptr, dim_size, dim, reduce='mean')
+        return self.weighted_reduce(x, index, weight, ptr, dim_size, dim,
+                                    reduce='sum')
 
 
 class MaxAggregation(Aggregation):